+++ example/lib/demo/simple/grid_view_demo.dart
@@ -3,7 +3,6 @@ import 'package:flutter/material.dart';
 
 import 'package:loading_more_list/loading_more_list.dart';
 
-
 import '../../data/tu_chong_repository.dart';
 import '../../data/tu_chong_source.dart';
 import '../../widget/item_builder.dart';
@@ -23,7 +22,7 @@ class GridViewDemo extends StatefulWidget {
 }
 
 class _GridViewDemoState extends State<GridViewDemo> {
- late TuChongRepository listSourceRepository;
+  late TuChongRepository listSourceRepository;
   @override
   void initState() {
     listSourceRepository = TuChongRepository();