+++ example/lib/demo/simple/sliver_grid_demo.dart
@@ -3,7 +3,6 @@ import 'package:flutter/material.dart';
 
 import 'package:loading_more_list/loading_more_list.dart';
 
-
 import '../../data/tu_chong_repository.dart';
 import '../../data/tu_chong_source.dart';
 import '../../widget/item_builder.dart';
@@ -23,7 +22,7 @@ class SliverGridDemo extends StatefulWidget {
 }
 
 class _SliverGridDemoState extends State<SliverGridDemo> {
- late TuChongRepository listSourceRepository;
+  late TuChongRepository listSourceRepository;
   @override
   void initState() {
     listSourceRepository = TuChongRepository();
@@ -40,7 +39,6 @@ class _SliverGridDemoState extends State<SliverGridDemo> {
   Widget build(BuildContext context) {
     return LoadingMoreCustomScrollView(
       showGlowLeading: false,
-      rebuildCustomScrollView: true,
       slivers: <Widget>[
         LoadingMoreSliverList<TuChongItem>(SliverListConfig<TuChongItem>(
           itemBuilder: itemBuilder,